omx{audio,video}{dec,enc}: sequentially disable ports because buffers are not shared
For the history, the parallel disable port has been introduced by:
"00be69f omxvideodec: Disable output port when setting a new format"
and then replicated to videoenc, audiodec and audioenc.

This is only required to do that if buffers are shared bettween ports.
But for decoders and encoders the input and output buffer are of
different nature by definition (bitstream vs images). So they cannot be
shared.

Also starting from IL 1.2.0 they now written in the spec that the
parallel disable is not allowed and will return an error. Except
for when buffer are shared.

Again here we know in advance that they are not shared so let's always
do a sequential disable.